Update to use same common components for steem trending charts
