+++ src/client/components/Sidebar/CryptoTrendingCharts.less
@@ -1,6 +1,6 @@
 @import (reference) "../../styles/custom.less";
 
-.SteemTrendingCharts {
+.CryptoTrendingCharts {
   background: @white;
   border-radius: 4px;
   border: 1px solid @white-gainsboro;
