+++ src/client/wallet/UserWalletTransactions.less
@@ -3,7 +3,6 @@
 
 .UserWalletTransactions {
   background-color: @white;
-  border: 1px solid @white-gainsboro;
   border-right: 0;
   border-left: 0;
   border-radius: 0;
@@ -105,7 +104,6 @@
   }
 
   @media @small {
-    border: 1px solid @white-gainsboro;
     border-radius: @border-radius-base;
   }
